splitting a project
